DermatologyQ&A Bot/requirements.txt
langchain
langchain-google-genai
python-dotenv
huggingface_hub
sentence-transformers
faiss-cpu
streamlit